Split local copy error definitions into dedicated module
